requirements.txt
streamlit
langchain-google-genai
pillow
dash
pandas
python-dotenv
dash-bootstrap-components
langchain
langchain-community
langchain-experimental
langchain-anthropic